+++ serializers/guest.js
@@ -1,9 +1,0 @@
-const JSONAPISerializer = require('jsonapi-serializer').Serializer;
-
-function Guest(data) {
-  return new JSONAPISerializer('guests', {
-    attributes: ['email'],
-  }).serialize(data);
-}
-
-module.exports = Guest;