+++ utils/messages.js
@@ -1,0 +1,7 @@
+module.exports = {
+  ERROR_UNEXPECTED: 'An unexpected error occurred. Please reach out for help in our Slack community or create a Github issue with following error:',
+  ERROR_MISSING_PROJECT_NAME: 'Missing project name in the command.',
+  HINT_MISSING_PROJECT_NAME: 'Please specify a project name. Type lumber help for more information.',
+  ERROR_NOT_PARSABLE_CONNECTION_URL: 'Cannot parse the database dialect. Please, check the syntax of the database connection string.',
+  HINT_DIRECTORY_ALREADY_EXISTS: 'Please retry with another project name.',
+};